Create EXERCISE 1.6: HELLO DOCKER HUB
Run docker run -it devopsdockeruh/pull_exercise.

It will wait for your input. Navigate through Docker hub to find the docs and Dockerfile that was used to create the image.

Read the Dockerfile and/or docs to learn what input will get the application to answer a "secret message".

Submit the secret message and command(s) given to get it as your answer.