Create EXERCISE 1.7: IMAGE FOR SCRIPT
Dockerfile for this excercise

Create a Dockerfile for a new image that starts from ubuntu:20.04 and add instructions to install curl into that image. Then add instructions to copy the script file into that image and finally set it to run on container start using CMD.

After you have filled the Dockerfile, build the image with the tag "curler".